% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{get_pinyin}
\alias{get_pinyin}
\title{Get pinyin of Chinese characters}
\usage{
get_pinyin(char)
}
\arguments{
\item{char}{A character vector of Chinese characters.}
}
\value{
A character vector with the same length as input `char`. The returned integer
vector is always named using input `char`.
}
\description{
`get_pinyin()` returns the pinyin of input Chinese characters. Both
Simplified Chinese and Traditional Chinese are supported.
}
\note{
Currently, polyphonics are not supported.
}
\examples{
get_pinyin(c("平安", "行"))

}
