% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{get_wuge}
\alias{get_wuge}
\title{Get WuGe information of Chinese names}
\usage{
get_wuge(xing, ming, to_trad = TRUE)
}
\arguments{
\item{xing}{A character vector of Chinese characters indicating the last
names.}

\item{ming}{A character vector of Chinese characters indicating the first
names.}

\item{to_trad}{If `TRUE`, input characters are converted to Traditional
Chinese using [sim2trad] before calculation. If [sim2trad] does not
give correct results, you may want to set `to_trad` to FALSE and
directly specify Tradition Chinese characters as input. Default:
`TRUE`}
}
\description{
`sim2tra()` uses the dictionary `STCharacter.txt` distributed via
[`OpenCC`](https://opencc.byvoid.com/) to convert Simplified Chinese to
Traditional Chinese.
}
\examples{
get_wuge(c("张", "李"), c("三", "四五"))

}
