% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\title{Calculate character stroke combinations that meet specified requirements}
\usage{
name(
  xing,
  num = 2L,
  min = NULL,
  max = NULL,
  fixed = NULL,
  shuli = "吉",
  sancai = shuli,
  common = TRUE
)
}
\arguments{
\item{xing}{A character vector of Chinese characters indicating the last
names.}

\item{num}{An integer indicating number of characters in the given names.
Default: `2L`.}

\item{min, max}{An integer indicating the minimum and maximum stroke numbers
in the given names. `NULL` means no restrictions. Default: `NULL`.}

\item{fixed}{An integer vector that has a length of `num` indicating the
fixed stroke number in the given names. `0` means do not fix. E.g. if
`num` is `2L`, `c(0, 10)` means that the second character in the
given names should has a stroke number of `10`, while there is no
fixed stroke number for the first character. Default: `NULL`.}

\item{shuli, sancai}{A string indicating the minimum luck for WuGe and
SanCai. Default: \code{"\u5409"}.}

\item{common}{If `TRUE`, only around 3500 common Chinese characters will be
used. Default: `TRUE`}
}
\value{
An `WuGeName` object.
}
\description{
Calculate character stroke combinations that meet specified requirements
}
