% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{sim2trad}
\alias{sim2trad}
\title{Convert Simplified Chinese to Traditional Chinese}
\usage{
sim2trad(char, all = FALSE)
}
\arguments{
\item{char}{A character vector of Simplified Chinese characters.}

\item{all}{If `TRUE`, all possible match results of Traditional Chinese
characters are returned. If `FALSE`, only the first match is returned.
Default: `FALSE`.}
}
\value{
If `all` is `FALSE`, a character vector with the same length as input `char`.
Otherwise, a list of character vector with the same length as input `char`.
The returned character vector or list is always named using input `char`.
}
\description{
`sim2trad()` uses the dictionary `STCharacter.txt` distributed via
[`OpenCC`](https://opencc.byvoid.com/) to convert Simplified Chinese to
Traditional Chinese.
}
\note{
Currently, {wuge} package did not use `OpenCC` library. Instead, a simple
character-level mapping is used.
}
\examples{
sim2trad(c("千万", "鼠标"))
sim2trad(c("千万", "鼠标"), all = TRUE)

}
