% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{stroke}
\alias{stroke}
\title{Get stroke number of Chinese characters}
\usage{
stroke(char)
}
\arguments{
\item{char}{A character vector of Chinese characters.}
}
\value{
An integer vector with the same length as input `char`.
The returned integer vector is always named using input `char`.
}
\description{
`get_stroke()` returns the stroke number of input Chinese characters. Both
Simplified Chinese and Traditional Chinese are supported.
}
\examples{
stroke(c("千万", "鼠标"))

}
