% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuge-package.R
\docType{package}
\name{wuge-package}
\alias{wuge}
\alias{wuge-package}
\title{wuge: Construct Chinese names based on 'WuGe' rules}
\description{
Provide utilities to construct Chinese names and evaluate them based on 'WuGe' rules.
}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})

}
\keyword{internal}
